Fix support for dashes in sample name
The mapped read count concatenation script was not prepared to handle sample names that contain dashes ('-'). It would just split file names by the last dash and use that as sample name. E.g. if you have sample names like "A-very-long-sample-name-1", it would return only the last part: "1". By itself that may not be such a big problem, but when the Snakemake rule `quantify_output` tries to merge the counts with the classifications based on sample name and scaffold name, this will fail and you will end up with a Sample_composition_graph.html file with 0 for all taxa.

TLDR:
In short, this patch solves the problem of faulty composition graphs for datasets in which sample names have dashes in them.