Fix bug: NA values for NR_ORFs
Draw heatmaps crashed on some samples, and it seemed that this was caused by 'non-finite' values being cast to integer values in `nr_orfs = df["Nr_ORFs"].astype(int)`.
This line of code is executed for taxa for which <=3 scaffolds are present. When more scaffolds are present, the heatmap will show 'aggregated statistics', and the dataframe for those statistics is filled with zeroes for NA values (see line 367: `new_df = new_df.fillna(0)`).
Therefore, I think a solution is to replace NA values in this piece of code as well.