Integrated the RA.smk to the QC and cleaning rules of the core workflow. The seperate sample_sheet.sh can now be removed. Also added a multiQC report, it's still a bit dirty since the names of files in the report are a bit wonky, see the TODO comments in the code. The wrapper still needs to be cleaned up to this new situation.
I found something weird/bug, the GC content track in IGVjs is no longer visible. Dito in the core workflow, this used to work... TODO for later.

I've also outcommented the read-length checker. Need to reactivate later. We also need to reduce the runtime of this script, takes too long.