Fixed the PCR duplicate removal for Jovian-RA. It no longer gives inflated coverage values. Duplicates are now hard removed instead of only marked.
By default, duplicates are now also removed in Jovian-meta. Duplicates are hard removed instead of only marked as DUP since not all downstream software is able to recognize and adapt to this flag.

For debugging purposes only, the dup removal for Jovian-meta can be turned off via a boolean in the pipeline_parameter file.

The hardcoded removal of PCR duplicates removes inflated coverage values (giving a false sense of confidence) while decreasing filesize and downstream processes' turnaround time.

Had to update samtools to 1.10 in scaffold_analyses.yaml for the rmDup functionality.