Create helper environment via mamba
I was helping our two Avans students (Bob van Rijthoven, Diemo Perre) get Jovian setup and they'd been stuck on conda resolution. On a previous installation it'd gone very well, but, we needed to reconfigure and reinstall and it stuck for days at a time on creating the Jovian_helper_env. I didn't see anything obvious slowing down the resolution, but on running a manual `conda env create -f ...` with `-vvv` it was just stuck in the SAT solver.

Creating this environment via mamba was instantaneous in comparison. It seems you're using mamba already in some of the per tool environments, I think it may make sense for the helper environment (and the main env possibly?) to be created via mamba as well.